Separate ops in operators/cv2 that do and do not require codecs so they're easier to include/exclude from a build. Remove jpeg2000 from opencv file formats. It costs 1MB and is (afaict) not a common format. Add ability to enable/disable cv2 ops to gen_selectedops.py.
