+++ cmake/externals/opencv.cmake
@@ -7,10 +7,10 @@ set(BUILD_IPP_IW                OFF CACHE INTERNAL "")
 set(BUILD_ITT                   OFF CACHE INTERNAL "")
 set(BUILD_JASPER                OFF CACHE INTERNAL "")
 set(BUILD_JAVA                  OFF CACHE INTERNAL "")
-# set(BUILD_JPEG                  OFF CACHE INTERNAL "")
+set(BUILD_JPEG                  OFF CACHE INTERNAL "")
 set(BUILD_OBJC                  OFF CACHE INTERNAL "")
-# set(BUILD_OPENJPEG              OFF CACHE INTERNAL "")
-# set(BUILD_PNG                   OFF CACHE INTERNAL "")
+set(BUILD_OPENJPEG              OFF CACHE INTERNAL "")
+set(BUILD_PNG                   OFF CACHE INTERNAL "")
 set(BUILD_opencv_apps           OFF CACHE INTERNAL "")
 set(BUILD_opencv_calib3d        OFF CACHE INTERNAL "")
 set(BUILD_opencv_dnn            OFF CACHE INTERNAL "")
@@ -18,7 +18,7 @@ set(BUILD_opencv_features2d     OFF CACHE INTERNAL "")
 set(BUILD_opencv_flann          OFF CACHE INTERNAL "")
 set(BUILD_opencv_gapi           OFF CACHE INTERNAL "")
 set(BUILD_opencv_highgui        OFF CACHE INTERNAL "")
-# set(BUILD_opencv_imgcodecs      OFF CACHE INTERNAL "")
+set(BUILD_opencv_imgcodecs      OFF CACHE INTERNAL "")
 set(BUILD_opencv_java           OFF CACHE INTERNAL "")
 set(BUILD_opencv_js             OFF CACHE INTERNAL "")
 set(BUILD_opencv_ml             OFF CACHE INTERNAL "")
@@ -46,6 +46,7 @@ set(WITH_CAP_IOS                OFF CACHE INTERNAL "")
 set(WITH_CAROTENE               OFF CACHE INTERNAL "")
 set(WITH_CLP                    OFF CACHE INTERNAL "")
 set(WITH_CPUFEATURES            OFF CACHE INTERNAL "")
+set(WITH_DIRECTX                OFF CACHE INTERNAL "")
 set(WITH_DSHOW                  OFF CACHE INTERNAL "")
 set(WITH_EIGEN                  OFF CACHE INTERNAL "")
 set(WITH_FFMPEG                 OFF CACHE INTERNAL "")
@@ -62,17 +63,17 @@ set(WITH_INF_ENGINE             OFF CACHE INTERNAL "")
 set(WITH_IPP                    OFF CACHE INTERNAL "")
 set(WITH_ITT                    OFF CACHE INTERNAL "")
 set(WITH_JASPER                 OFF CACHE INTERNAL "")
-# set(WITH_JPEG                   OFF CACHE INTERNAL "")
+set(WITH_JPEG                   OFF CACHE INTERNAL "")
 set(WITH_MSMF                   OFF CACHE INTERNAL "")
 set(WITH_NGRAPH                 OFF CACHE INTERNAL "")
 set(WITH_ONNX                   OFF CACHE INTERNAL "")
 set(WITH_OPENCL                 OFF CACHE INTERNAL "")
 set(WITH_OPENCL_SVM             OFF CACHE INTERNAL "")
 set(WITH_OPENEXR                OFF CACHE INTERNAL "")
-# set(WITH_OPENJPEG               OFF CACHE INTERNAL "")
+set(WITH_OPENJPEG               OFF CACHE INTERNAL "")
 set(WITH_OPENMP                 OFF CACHE INTERNAL "")
 set(WITH_OPENVX                 OFF CACHE INTERNAL "")
-# set(WITH_PNG                    OFF CACHE INTERNAL "")
+set(WITH_PNG                    OFF CACHE INTERNAL "")
 set(WITH_PROTOBUF               OFF CACHE INTERNAL "")
 set(WITH_PTHREADS_PF            OFF CACHE INTERNAL "")
 set(WITH_QUIRC                  OFF CACHE INTERNAL "")
@@ -88,11 +89,9 @@ if (OCOS_ENABLE_OPENCV_CODECS)
   set(BUILD_opencv_imgcodecs  ON CACHE INTERNAL "")
 
   set(BUILD_JPEG              ON CACHE INTERNAL "")
-  set(BUILD_OPENJPEG          ON CACHE INTERNAL "")
   set(BUILD_PNG               ON CACHE INTERNAL "")
 
   set(WITH_JPEG               ON CACHE INTERNAL "")
-  set(WITH_OPENJPEG           ON CACHE INTERNAL "")
   set(WITH_PNG                ON CACHE INTERNAL "")
 endif()
 
@@ -121,7 +120,6 @@ list(APPEND opencv_INCLUDE_DIRS ${OPENCV_CONFIG_FILE_INCLUDE_DIR})
 list(APPEND opencv_INCLUDE_DIRS
     ${OPENCV_MODULE_opencv_core_LOCATION}/include
     ${OPENCV_MODULE_opencv_imgproc_LOCATION}/include)
-
 set(opencv_LIBS "")
 list(APPEND opencv_LIBS opencv_core opencv_imgproc)
 
@@ -141,3 +139,4 @@ if (CMAKE_SYSTEM_NAME MATCHES "Windows")
         set_target_properties(${p} PROPERTIES FOLDER "externals/opencv")
     endforeach()
 endif()
+
